[refactor] move PositionIterator to its own file
